Moving test files (2)
