--- src/main/scala/com/wavesplatform/AutoShutdown.scala
@@ -1,13 +1,0 @@
-package com.wavesplatform
-
-import scala.concurrent.Future
-
-trait AutoShutdown {
-  def shutdown(): Future[Unit]
-}
-
-object AutoShutdown {
-  object Done extends AutoShutdown {
-    override def shutdown(): Future[Unit] = Future.unit
-  }
-}
