+++ src/test/scala/units/BlockBriefValidationTestSuite.scala
@@ -1,6 +1,5 @@
 package units
 
-import com.wavesplatform.extensions.ExtensionDomain
 import com.wavesplatform.wallet.Wallet
 import units.ELUpdater.State.Working
